query_system/requirements.txt
streamlit
groq
openai
langsmith
langchain-core
langgraph
duckdb
pandas
python-dotenv
